wip: advance

+++ unity-renderer/Assets/DCLPlugins/ECS7/TestsUtils/Components/UI/ECS7TestUIUtils.cs
@@ -8,7 +8,7 @@ namespace DCLPlugins.ECS7.TestsUtils
         public static PBUiTransform CreatePBUiTransformDefaultModel()
         {
             var model = new PBUiTransform();
-            model.Parent = SpecialEntityId.SCENE_ROOT_ENTITY;
+            model.Parent = (int)SpecialEntityId.SCENE_ROOT_ENTITY;
             model.AlignContent = YGAlign.FlexStart;
             model.AlignSelf = YGAlign.Auto;
             model.AlignItems = YGAlign.Auto;
